chores: rename Recent History component and add key in PlayerItem
